Revert "remove unneded plugins"
This reverts commit 496cf333eb8d5952f6250df706124c61bf9cb575.
