+++ lua/plugins/treesitter.lua
@@ -1,0 +1,1 @@
+return {}
